fix: Call setParents on `MemDB` tables (#217)
I noticed the [table list for Bitbucket plugin is broken in the Hub](https://hub.cloudquery.io/plugins/source/cloudquery/bitbucket/latest/tables). The reason is that we don't call `setParents` thus publish all tables as top level ones. The UI then tries to render items with duplicate keys and fails.

This PR fixes the MemDB plugin in the SDK which is used as a reference for other plugins. I'll fix the Bitbucket plugin in a separate PR.

There might be a better way to implicitly initialize the parents but I don't we should spend more time on it